requirements.txt
Pillow==6.2.0
psycopg2==2.8.3
psycopg2-binary==2.8.6
Django==3.1.4
django-allauth==0.43.0
django-crispy-forms==1.10.0
django-debug-toolbar==3.2
django-extensions==3.0.9
django-filter==2.4.0
whitenoise==5.2.0